Optimize testing coverage and refactor code

+++ src/main/java/com/hhaouari/roverscan/entities/Plateau.java
@@ -6,10 +6,10 @@ import java.util.Objects;
 
 @Data
 public class Plateau {
-    private int height;
-    private int width;
+    private long height;
+    private long width;
 
-    public Plateau(int width, int height) {
+    public Plateau(long width, long height) {
         this.width = width;
         this.height = height;
     }
