+++ src/main/java/com/hhaouari/roverscan/entities/Rover.java
@@ -7,15 +7,15 @@ import java.util.Objects;
 
 @Data
 public class Rover {
-    private int x;
-    private int y;
+    private long x;
+    private long y;
     private Direction direction;
     private String instructions;
 
     public Rover() {
     }
 
-    public Rover(int x, int y, Direction direction, String instructions) {
+    public Rover(long x, long y, Direction direction, String instructions) {
         this.x = x;
         this.y = y;
         this.direction = direction;
